api/pyproject.toml
[project]
name = "bl-intelligence-api"
version = "0.1.0"
description = "BL/GL Intelligence Platform — FastAPI + Claude Agent SDK service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "claude-agent-sdk>=0.1.2",
  "anthropic>=0.40.0",

  # Persistence
  "supabase>=2.15.1",
  "asyncpg>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.36",

  # BigQuery / Cortex
  "google-cloud-bigquery>=3.25.0",
  "google-auth>=2.23.0",

  # HTTP / SSE
  "httpx>=0.27.0",
  "sse-starlette>=2.1.3",

  # Auth
  "pyjwt[crypto]>=2.9.0",

  # Misc
  "python-dotenv>=1.0.1",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "hypothesis>=6.112.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
